Bug fix and routing description added
During walkthrough with BSC staff, a bug was discovered. From the East touch panel in combined mode, when a source was routed to the West projector, it wouldn't power on. Ant recommended that a description should be added below the source selection that indicates which display was being routed to.